+++ evalml/tests/model_understanding_tests/prediction_explanations_tests/test_explainers.py
@@ -1,5 +1,4 @@
 import json
-from collections import OrderedDict
 from unittest.mock import MagicMock, patch
 
 import numpy as np
@@ -10,7 +9,6 @@ import woodwork as ww
 from evalml.exceptions import PipelineScoreError
 from evalml.model_understanding.prediction_explanations.explainers import (
     abs_error,
-    clean_format_tree,
     cross_entropy,
     explain_prediction,
     explain_predictions,
@@ -621,32 +619,3 @@ def test_json_serialization(problem_type, X_y_regression, linear_regression_pipe
 
     report = explain_predictions(pipeline, pd.DataFrame(X[:1]), output_format="dict")
     assert json.loads(json.dumps(report)) == report
-
-
-def test_clean_format_tree(tree_estimators, logit_estimator):
-    est_class, est_reg = tree_estimators
-    est_logit = logit_estimator
-
-    with pytest.raises(ValueError, match="Tree structure reformatting is not supported for non-Tree estimators"):
-        clean_format_tree(est_logit)
-
-    formatted_ = clean_format_tree(est_reg)
-    tree_ = est_reg._component_obj.tree_
-
-    assert isinstance(formatted_, OrderedDict)
-    assert formatted_['Feature'] == tree_.feature[0]
-    assert formatted_['Threshold'] == tree_.threshold[0]
-    assert all([a == b for a, b in zip(formatted_['Value'][0], tree_.value[0][0])])
-    left_child_name_ = formatted_['Left_Child']['Name']
-    right_child_name_ = formatted_['Right_Child']['Name']
-    left_child_threshold_ = formatted_['Left_Child']['Threshold']
-    right_child_threshold_ = formatted_['Right_Child']['Threshold']
-    left_child_value_ = formatted_['Left_Child']['Value']
-    right_child_value_ = formatted_['Right_Child']['Value']
-    assert left_child_name_[left_child_name_.index('_') + 1:] == str(tree_.children_left[0])
-    assert right_child_name_[right_child_name_.index('_') + 1:] == str(tree_.children_right[0])
-    assert left_child_threshold_ == tree_.threshold[tree_.children_left[0]]
-    assert right_child_threshold_ == tree_.threshold[tree_.children_right[0]]
-    # Check that the immediate left and right child of the root node have the correct values
-    assert all([a == b for a, b in zip(left_child_value_[0], tree_.value[tree_.children_left[0]][0])])
-    assert all([a == b for a, b in zip(right_child_value_[0], tree_.value[tree_.children_right[0]][0])])
